requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-DebugToolbar==0.15.1
forex-python==1.8
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
python-dotenv==1.0.1
requests==2.32.3
simplejson==3.19.3
urllib3==2.2.2
Werkzeug==3.0.4
